# frontend/shared/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "DeviceConfigAuditLogItem": {
        "additionalProperties": false,
        "properties": {
          "actorId": {
            "format": "int64",
            "type": "integer"
          },
          "actorType": {
            "type": "string"
          },
          "commandId": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "deviceId": {
            "format": "int64",
            "type": "integer"
          },
          "eventType": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "revision": {
            "format": "int64",
            "type": "integer"
          },
          "summary": {
            "type": "string"
          }
        },
        "required": [
          "actorId",
          "actorType",
          "createdAt",
          "deviceId",
          "eventType",
          "id",
          "revision",
          "summary"
        ],
        "type": "object"
      },
      "DeviceConfigAuditLogsResponse": {
        "additionalProperties": false,
        "properties": {
          "limit": {
            "format": "int64",
            "type": "integer"
          },
          "logs": {
            "items": {
              "$ref": "#/components/schemas/DeviceConfigAuditLogItem"
            },
            "type": "array"
          },
          "offset": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "limit",
          "logs",
          "offset"
        ],
        "type": "object"
      },
      "DeviceConfigCommandItem": {
        "additionalProperties": false,
        "properties": {
          "actorId": {
            "format": "int64",
            "type": "integer"
          },
          "actorType": {
            "type": "string"
          },
          "appliedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "baseRevision": {
            "format": "int64",
            "type": "integer"
          },
          "createdAt": {
            "type": "string"
          },
          "failureReason": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "mutation": {
            "$ref": "#/components/schemas/JsonObject"
          },
          "status": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "targetRevision": {
            "format": "int64",
            "type": "integer"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "actorId",
          "actorType",
          "baseRevision",
          "createdAt",
          "id",
          "mutation",
          "status",
          "summary",
          "targetRevision",
          "updatedAt"
        ],
        "type": "object"
      },
      "DeviceConfigCommandRequest": {
        "additionalProperties": false,
        "properties": {
          "baseRevision": {
            "format": "int64",
            "type": "integer"
          },
          "mutation": {
            "$ref": "#/components/schemas/JsonObject"
          },
          "summary": {
            "type": "string"
          }
        },
        "required": [
          "baseRevision",
          "mutation",
          "summary"
        ],
        "type": "object"
      },
      "DeviceConfigStateResponse": {
        "additionalProperties": false,
        "properties": {
          "deviceId": {
            "format": "int64",
            "type": "integer"
          },
          "mirrorContent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JsonObject"
              },
              {
                "type": "null"
              }
            ]
          },
          "pendingCommands": {
            "items": {
              "$ref": "#/components/schemas/DeviceConfigCommandItem"
            },
            "type": "array"
          },
          "revision": {
            "format": "int64",
            "type": "integer"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "deviceId",
          "pendingCommands",
          "revision",
          "updatedAt"
        ],
        "type": "object"
      },
      "DeviceItem": {
        "additionalProperties": false,
        "properties": {
          "appVersion": {
            "type": "string"
          },
          "capabilities": {
            "$ref": "#/components/schemas/JsonObject"
          },
          "createdAt": {
            "type": "string"
          },
          "deviceModel": {
            "type": "string"
          },
          "deviceName": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "lastSeenAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "localAddresses": {
            "$ref": "#/components/schemas/JsonObject"
          },
          "platform": {
            "type": "string"
          },
          "revokedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "updatedAt": {
            "type": "string"
          },
          "userId": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "appVersion",
          "capabilities",
          "createdAt",
          "deviceModel",
          "deviceName",
          "displayName",
          "enabled",
          "id",
          "localAddresses",
          "platform",
          "updatedAt",
          "userId"
        ],
        "type": "object"
      },
...
      "JsonObject": {
        "additionalProperties": true,
        "type": "object"
      },
...
      "PatchDeviceRequest": {
        "additionalProperties": false,
        "properties": {
          "displayName": {
            "type": [
              "string",
              "null"
            ]
          },
          "enabled": {
            "type": [
              "boolean",
              "null"
            ]
          }
        },
        "type": "object"
      },
...
      "SimpleOKResponse": {
        "additionalProperties": false,
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Contract for the Xinyi Relay backend, Android agent, Web console, and Desktop console.",
    "title": "Xinyi Relay Remote API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/devices/{id}": {
      "patch": {
        "operationId": "patchDevice",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchDeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceItem"
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/devices/{id}/config": {
      "get": {
        "operationId": "getDeviceConfig",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceConfigStateResponse"
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/devices/{id}/config/audit": {
      "get": {
        "operationId": "listDeviceConfigAuditLogs",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceConfigAuditLogsResponse"
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/devices/{id}/config/commands": {
      "post": {
        "operationId": "queueDeviceConfigCommand",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceConfigCommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceConfigCommandItem"
                }
              }
            },
            "description": "Created"
          }
        }
      }
    },
    "/api/v1/devices/{id}/revoke": {
      "post": {
        "operationId": "revokeDevice",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleOKResponse"
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "description": "Same-origin backend",
      "url": "/"
    }
  ]
}